OO-4287: split the queries for admin. with not members included
In assessment tool, split the queries to retrieve the assessed
identities between the participants and not course members to speed up
loading of the assessment tool (overview and every node)

+++ src/main/java/org/olat/modules/assessment/ui/AssessmentStatisticsSmallController.java
@@ -93,7 +93,7 @@ public class AssessmentStatisticsSmallController extends BasicController {
 		numOfPassed = stats.getCountPassed();
 		mainVC.contextPut("numOfPassed", numOfPassed);
 		
-		int total = membersStatistics.getTotal();
+		float total = membersStatistics.getTotal();
 		int percentPassed = total <= 0 ? 0 : Math.round(100.0f * (stats.getCountPassed() / total));
 		mainVC.contextPut("percentPassed", percentPassed);
 		numOfFailed = stats.getCountFailed();
